🚀 Vercel 마이그레이션 및 배포 완료
- Netlify에서 Vercel로 플랫폼 마이그레이션
- API 엔드포인트를 Vercel Functions 형식으로 변환
- TypeScript 빌드 설정 최적화
- 배포 URL: https://hrfco-mcp-npwialnrt-kewns-projects.vercel.app
- QA 테스트 통과 및 모든 기능 정상 작동 확인
