+++ .gitignore
@@ -46,3 +46,5 @@ coverage/
 *.tmp
 *.temp
 .cache/
+.vercel
+.aider*
