+++ api/health.ts
@@ -1,23 +1,25 @@
-import type { Handler } from '@netlify/functions';
+import type { VercelRequest, VercelResponse } from '@vercel/node';
 
-const health: Handler = async (event) => {
-  const healthData = {
+export default async function health(req: VercelRequest, res: VercelResponse) {
+  const headers = {
+    'Content-Type': 'application/json',
+    'Access-Control-Allow-Origin': '*',
+  };
+
+  res.setHeader('Content-Type', 'application/json');
+  res.setHeader('Access-Control-Allow-Origin', '*');
+
+  res.status(200).json({
     status: 'healthy',
     timestamp: new Date().toISOString(),
+    service: 'HRFCO MCP Server',
     version: '1.0.0',
-    runtime: 'netlify',
-    api_keys: {
-      hrfco: !!process.env.HRFCO_API_KEY,
-    },
-  };
-
-  return {
-    statusCode: 200,
-    headers: {
-      'Content-Type': 'application/json',
-    },
-    body: JSON.stringify(healthData),
-  };
-};
-
-export default health;
+    environment: 'vercel',
+    features: [
+      '통합 검색 기능 (get_water_info)',
+      'ChatGPT 무한 반복 호출 방지',
+      '실시간 수위 데이터 조회',
+      '관측소 코드 매핑'
+    ]
+  });
+}
