+++ public/index.html
@@ -1,0 +1,1 @@
+HRFCO MCP Server
